+++ index.ts
@@ -1,1 +1,4 @@
-export * from './lib'
+export * from './readAs'
+export * from './readAs_ArrayBuffer'
+export * from './readAs_DataURL'
+export * from './readAs_Text'