+++ readAs_ArrayBuffer.d.ts
@@ -1,0 +1,2 @@
+export declare function readAs_ArrayBuffer(file: any): Promise<unknown>;
+export { readAs_ArrayBuffer as readAsArrayBuffer, };