+++ readAs_ArrayBuffer.ts
@@ -1,0 +1,7 @@
+import { readAs } from './readAs'
+export function readAs_ArrayBuffer(file) {
+	return readAs(file, 'ArrayBuffer')
+}
+export {
+	readAs_ArrayBuffer as readAsArrayBuffer,
+}