+++ readAs_DataURL.d.ts
@@ -1,0 +1,2 @@
+export declare function readAs_DataURL(file: any): Promise<unknown>;
+export { readAs_DataURL as readAsDataURL, };