+++ readAs_DataURL.ts
@@ -1,0 +1,7 @@
+import { readAs } from './readAs'
+export function readAs_DataURL(file) {
+	return readAs(file, 'DataURL')
+}
+export {
+	readAs_DataURL as readAsDataURL,
+}