+++ readAs_Text.d.ts
@@ -1,0 +1,2 @@
+export declare function readAs_Text(file: any): Promise<unknown>;
+export { readAs_Text as readAsText, };