+++ readAs_Text.ts
@@ -1,0 +1,7 @@
+import { readAs } from './readAs'
+export function readAs_Text(file) {
+	return readAs(file, 'Text')
+}
+export {
+	readAs_Text as readAsText,
+}